policy_analysis/pyproject.toml
[project]
name = "policy-analysis"
version = "0.1.0"
description = "Agentic RAG + policy analysis pipelines"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "pandas>=2.3.3",
    "faiss-cpu>=1.13.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "matplotlib>=3.10.7",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "dspy>=3.0.4",
    "hf-xet>=1.2.0",
]

[project.optional-dependencies]

dspy = [
    "datetime>=6.0",
    "dspy>=3.0.4",
]

agentic = [
    "sqlalchemy>=2.0.29,<3",
    "sqlmodel>=0.0.16,<0.1",
    "psycopg2-binary>=2.8.1,<2.9.10",
    "bertopic>=0.17.3",
    "nltk>=3.9.2",
    "matplotlib>=3.10.7",
    "wordcloud>=1.9.4",
    "qdrant-client>=1.15.1",
    "optuna>=4.5.0",
    "kaleido>=1.2.0",
    "together>=1.5.30",
    "scikit-learn>=1.6.1",
    "pymupdf4llm>=0.1.8",
    "plotly>=6.3.1",
]

all = [
    "policy-analysis[dspy,agentic]"
]

clustering = [
    "hdbscan>=0.8.40",
    "scikit-learn>=1.6.1",
    "umap-learn>=0.5.9.post2",
]



